\name{Example.3.1.4}
\alias{Example.3.1.4}
\docType{data}
\title{
One Sample Mean Equality
}
\description{
The test is finding whether there is a difference between the mean response of the test xbar and the reference value mu0

}
\usage{Example.3.1.4}

\references{
Chow SC, Shao J, Wang H. Sample Size Calculation in Clinical Research. New York: Marcel Dekker, 2003
}

\examples{
data(Example.3.1.4)
## maybe str(Example.3.1.4) ; plot(Example.3.1.4) ...
}
\keyword{datasets}
