\name{TrialSize-package}
\alias{TrialSize-package}
\alias{TrialSize}
\docType{package}
\title{
Sample Size calculation in Clinical Research
}
\description{
More than 80 functions in this package are widely used to calculate sample size in clinical trial research studies.

This package covers the functions in Chapter 3,4,6,7,9,10,11,12,14,15 of the reference book. 
}
\details{
\tabular{ll}{
Package: \tab TrialSize\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2013-05-31\cr
License: \tab GPL ( >=2  \cr
LazyLoad: \tab yes\cr
}

}

\author{
author: Ed Zhang <ed.zhang.jr@gmail.com>

        Vicky Qian Wu <wuqian7@gmail.com>

        Harry G. Zhang (Quality check)
   
        Shein-Chung Chow

maintainer: Vicky Qian Wu <wuqian7@gmail.com> 
}
\references{
Chow SC, Shao J, Wang H. Sample Size Calculation in Clinical Research. New York: Marcel Dekker, 2008
}
\keyword{ package }
