\name{TrialSize-package}
\alias{TrialSize-package}
\alias{TrialSize}
\docType{package}
\title{
Sample Size calculation in Clinical Research
}
\description{
The functions used to calculate the sample size in most of the clinical trial and research studies.

Here it covers the functions in Chapter 3,4,6,7,9,10,11,12,14,15 of the reference book. 
}
\details{
\tabular{ll}{
Package: \tab TrialSize\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-12-09\cr
License: \tab GPL ( >=2  \cr
LazyLoad: \tab yes\cr
}

}

\author{
author: Ed Zhang <ed.zhang.jr@gmail.com>
   
        Shein-Chung Chow

maintainer: Qian Wu <qwu@acr.org>
}
\references{
Chow SC, Shao J, Wang H. Sample Size Calculation in Clinical Research. New York: Marcel Dekker, 2008
}
\keyword{ package }
